requirements.txt
wandb~=0.16.3
matplotlib~=3.8.2
omegaconf~=2.3.0
torchmetrics~=1.2.1
pyyaml~=6.0.1
numpy~=1.26.3
albumentations~=1.3.1
pillow~=9.5.0
tqdm~=4.66.1
torchvision~=0.17.0
hydra-core~=1.3.2